requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.2.3
et_xmlfile==2.0.0
fuzzywuzzy==0.18.0
Levenshtein==0.27.1
lxml==5.4.0
numpy==2.3.0
openpyxl==3.1.5
pandas==2.3.0
pillow==11.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-Levenshtein==0.27.1
pytz==2025.2
pywin32==310
RapidFuzz==3.13.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2 